+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: imedgar <imedgar@21-school.ru>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/07/17 20:57:47 by imedgar           #+#    #+#             */
-/*   Updated: 2020/07/17 21:33:48 by imedgar          ###   ########.fr       */
+/*   Updated: 2020/07/18 19:10:49 by imedgar          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ int				ft_putchars(char c, int nbr, char *str, char pdn);
 int				ft_print_char(char c, t_fl *fl);
 int				ft_print_str(char *str, t_fl *fl);
 int				ft_print_adr(long long num, t_fl *fl);
-int				ft_print_nbr(long adr, t_fl *fl, char base, char x);
+int				ft_print_nbr(size_t adr, t_fl *fl, char base, char x);
 void			ft_init(t_fl *fl);
 long long		ft_star_width(long long star, t_fl *fl);
